@@ -1,14 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.jagerdev.foxhoundpricetracker"
         minSdkVersion 19
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
         vectorDrawables.useSupportLibrary = true
     }
@@ -28,25 +28,27 @@ android {
 }
 
 repositories {
+    google()
     maven { url 'https://jitpack.io' }
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
-    implementation 'com.android.support:design:26.1.0'
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    implementation 'com.android.support:support-v4:26.1.0'
-    implementation 'com.android.support:support-vector-drawable:26.1.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.vectordrawable:vectordrawable:1.0.0'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    compile group: 'org.xerial', name: 'sqlite-jdbc', version: '3.21.0'
-    compile group: 'org.nanohttpd', name: 'nanohttpd', version: '2.3.0'
+    androidTestImplementation 'androidx.test:runner:1.1.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
+    implementation group: 'org.xerial', name: 'sqlite-jdbc', version: '3.21.0'
+    implementation group: 'org.nanohttpd', name: 'nanohttpd', version: '2.3.1'
     implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0-alpha'
     api 'com.google.guava:guava:27.1-android'
-    compile('jager.pricetracker.engine:PriceTrackerEngine:1.0.0')
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation('jager.pricetracker.engine:PriceTrackerEngine:1.0.0')
             {
                 exclude module: 'org.restlet.ext.json'
                 exclude group: 'org.restlet.jee'
