@@ -1,7 +1,7 @@
 package com.jagerdev.foxhoundpricetracker;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.jagerdev.foxhoundpricetracker.utils.ServiceRunHandler;
 
