@@ -2,9 +2,9 @@ package com.jagerdev.foxhoundpricetracker.products;
 
 import android.content.Context;
 import android.graphics.Color;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.content.ContextCompat;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.core.content.ContextCompat;
 import android.text.util.Linkify;
 import android.view.LayoutInflater;
 import android.view.View;
